refactor(runtime): extract github api client module (#1278)
